fix(compile): fixed duplicates issue and combined services and job funcs
